+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,10 +3,12 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Fixed Invalid and Incomplete wdio test logic to ensure states are off after tests runs.
+
 ### Changed
 * Resolved lint warnings for multiple empty lines
 * Regenerate wdio screenshots
-* Fixed some wdio test logic
 
 4.18.0 - (October 30, 2019)
 ------------------